Changed return type of serial_read_byte(), removed debugging comments.
Fixed the return type of serial_read_byte() to be of type int, instead of
uint16. This prevented a return value of -1 to denote no data avaiable.
Also removed several printk statements specifying the systemcall occuring

+++ include/arch/console.h
@@ -24,6 +24,6 @@ int cons_getc(void);
 void kbd_intr(void); // irq 1
 void serial_intr(void); // irq 4
 void serial_send_byte(uint8_t b);
-uint8_t serial_read_byte();
+int serial_read_byte();
 
 #endif /* _CONSOLE_H_ */
